+++ R/convert_na_to.R
@@ -145,7 +145,7 @@ convert_na_to.character <- function(x, replacement = NULL, verbose = TRUE, ...) 
 #' @param replace_num Value to replace `NA` when variable is of type numeric.
 #' @param replace_char Value to replace `NA` when variable is of type character.
 #' @param replace_fac Value to replace `NA` when variable is of type factor.
-#' @inheritParams find_columns
+#' @inheritParams extract_column_names
 #'
 #' @rdname convert_na_to
 #' @export
@@ -160,10 +160,10 @@ convert_na_to.data.frame <- function(x,
                                      regex = FALSE,
                                      verbose = TRUE,
                                      ...) {
-  data <- x
+  my_data <- x
   select_nse <- .select_nse(
     select,
-    data,
+    data = my_data,
     exclude = exclude,
     ignore_case,
     regex = regex,