+++ R/select_nse.R
@@ -149,9 +149,8 @@
 # 3 types of symbols:
 # - unquoted variables
 # - objects that need to be evaluated, e.g data_find(iris, i) where
-#   i is a
-#   function arg or is defined before. This can also be a vector of names or
-#   positions.
+#   i is a function arg or is defined before. This can also be a
+#   vector of names or positions.
 # - functions (without parenthesis)
 
 # The first case is easy to deal with.