+++ tests/testthat/test-extract_column_names.R
@@ -1,83 +1,83 @@
-test_that("find_columns works as expected", {
+test_that("extract_column_names works as expected", {
   expect_identical(
-    find_columns(iris, starts_with("Sepal")),
+    extract_column_names(iris, starts_with("Sepal")),
     c("Sepal.Length", "Sepal.Width")
   )
 
   expect_identical(
-    find_columns(iris, starts_with("Sepal", "Petal")),
+    extract_column_names(iris, starts_with("Sepal", "Petal")),
     c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")
   )
 
   expect_identical(
-    find_columns(iris, ends_with("Width")),
+    extract_column_names(iris, ends_with("Width")),
     c("Sepal.Width", "Petal.Width")
   )
 
   expect_identical(
-    find_columns(iris, ends_with("Length", "Width")),
+    extract_column_names(iris, ends_with("Length", "Width")),
     c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")
   )
 
   expect_identical(
-    find_columns(iris, regex("\\.")),
+    extract_column_names(iris, regex("\\.")),
     c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")
   )
 
   expect_identical(
-    find_columns(iris, c("Petal.Width", "Sepal.Length")),
+    extract_column_names(iris, c("Petal.Width", "Sepal.Length")),
     c("Petal.Width", "Sepal.Length")
   )
 
   expect_identical(
-    find_columns(iris, contains("Wid")),
+    extract_column_names(iris, contains("Wid")),
     c("Sepal.Width", "Petal.Width")
   )
 
   expect_identical(
-    find_columns(iris, contains("en", "idt")),
+    extract_column_names(iris, contains("en", "idt")),
     c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")
   )
 
   expect_identical(
-    find_columns(mtcars, c("am", "gear", "cyl")),
+    extract_column_names(mtcars, c("am", "gear", "cyl")),
     c("am", "gear", "cyl")
   )
 
   expect_identical(
-    find_columns(mtcars, c("vam", "gear", "cyl")),
+    extract_column_names(mtcars, c("vam", "gear", "cyl")),
     c("gear", "cyl")
   )
 
-  expect_warning(expect_null(find_columns(mtcars, ends_with("abc"))))
+  expect_warning(expect_null(extract_column_names(mtcars, ends_with("abc"))))
 
   expect_identical(
-    find_columns(mtcars, regex("rb$")),
+    extract_column_names(mtcars, regex("rb$")),
     "carb"
   )
 
   expect_identical(
-    find_columns(mtcars, regex("^c")),
+    extract_column_names(mtcars, regex("^c")),
     c("cyl", "carb")
   )
 
-  expect_warning(expect_null(find_columns(mtcars, "^c")))
+  expect_warning(expect_null(extract_column_names(mtcars, "^c")))
 
   expect_identical(
-    find_columns(mtcars, regex("^C"), ignore_case = TRUE),
+    extract_column_names(mtcars, regex("^C"), ignore_case = TRUE),
     c("cyl", "carb")
   )
 
   expect_identical(
-    find_columns(iris, "Width$", regex = TRUE),
+    extract_column_names(iris, "Width$", regex = TRUE),
     c("Sepal.Width", "Petal.Width")
   )
 })
 
 
-test_that("find_columns from other functions", {
+test_that("extract_column_names from other functions", {
   test_fun1 <- function(data, i) {
-    find_columns(data, select = i)
+    extract_column_names(data, select = i)
   }
   expect_identical(
     test_fun1(iris, c("Sepal.Length", "Sepal.Width")),
@@ -90,7 +90,7 @@ test_that("find_columns from other functions", {
   )
 
   test_fun1a <- function(data, i) {
-    find_columns(data, select = i, regex = TRUE)
+    extract_column_names(data, select = i, regex = TRUE)
   }
   expect_identical(
     test_fun1a(iris, "Sep"),
@@ -98,7 +98,7 @@ test_that("find_columns from other functions", {
   )
 
   test_fun1b <- function(data, i) {
-    find_columns(data, select = i, regex = TRUE)
+    extract_column_names(data, select = i, regex = TRUE)
   }
   expect_identical(
     test_fun1b(iris, "Width$"),
@@ -106,7 +106,7 @@ test_that("find_columns from other functions", {
   )
 
   test_fun2 <- function(data) {
-    find_columns(data, select = starts_with("Sep"))
+    extract_column_names(data, select = starts_with("Sep"))
   }
   expect_identical(
     test_fun2(iris),
@@ -115,7 +115,7 @@ test_that("find_columns from other functions", {
 
   test_fun3 <- function(data) {
     i <- "Sep"
-    find_columns(data, select = starts_with(i))
+    extract_column_names(data, select = starts_with(i))
   }
   expect_identical(
     test_fun3(iris),
@@ -123,68 +123,68 @@ test_that("find_columns from other functions", {
   )
 })
 
-test_that("find_columns regex", {
+test_that("extract_column_names regex", {
   expect_identical(
-    find_columns(mtcars, select = "pg", regex = TRUE),
-    find_columns(mtcars, select = "mpg")
+    extract_column_names(mtcars, select = "pg", regex = TRUE),
+    extract_column_names(mtcars, select = "mpg")
   )
 })
 
-test_that("find_columns works correctly with minus sign", {
+test_that("extract_column_names works correctly with minus sign", {
   expect_identical(
-    find_columns(iris, -"Sepal.Length"),
+    extract_column_names(iris, -"Sepal.Length"),
     c("Sepal.Width", "Petal.Length", "Petal.Width", "Species")
   )
 
   expect_identical(
-    find_columns(iris, -c("Sepal.Length", "Petal.Width")),
+    extract_column_names(iris, -c("Sepal.Length", "Petal.Width")),
     c("Sepal.Width", "Petal.Length", "Species")
   )
 
   expect_identical(
-    find_columns(iris, -1),
+    extract_column_names(iris, -1),
     c("Sepal.Width", "Petal.Length", "Petal.Width", "Species")
   )
 
   expect_error(
-    find_columns(iris, -1:2),
+    extract_column_names(iris, -1:2),
     regexp = "can't mix negative"
   )
 
   expect_identical(
-    find_columns(iris, -(1:2)),
+    extract_column_names(iris, -(1:2)),
     c("Petal.Length", "Petal.Width", "Species")
   )
 
   expect_identical(
-    find_columns(iris, -c(1, 3)),
+    extract_column_names(iris, -c(1, 3)),
     c("Sepal.Width", "Petal.Width", "Species")
   )
 
   expect_identical(
-    find_columns(iris, -starts_with("Sepal", "Petal")),
+    extract_column_names(iris, -starts_with("Sepal", "Petal")),
     "Species"
   )
 
   expect_identical(
-    find_columns(iris, -ends_with("Length", "Width")),
+    extract_column_names(iris, -ends_with("Length", "Width")),
     "Species"
   )
 
   expect_identical(
-    find_columns(iris, -contains("en", "idt")),
+    extract_column_names(iris, -contains("en", "idt")),
     "Species"
   )
 
   expect_identical(
-    find_columns(iris, -c("Sepal.Length", "Petal.Width"), exclude = "Species"),
+    extract_column_names(iris, -c("Sepal.Length", "Petal.Width"), exclude = "Species"),
     c("Sepal.Width", "Petal.Length")
   )
 })
 
-test_that("find_columns with square brackets", {
+test_that("extract_column_names with square brackets", {
   expect_identical(
-    find_columns(mtcars, select = names(mtcars)[-1]),
-    find_columns(mtcars, select = 2:11)
+    extract_column_names(mtcars, select = names(mtcars)[-1]),
+    extract_column_names(mtcars, select = 2:11)
   )
 })